mm: fix the use of ALLOC_CMA
Movable allocations are allowed to allocate from CMA areas
only when __GFP_CMA is set in the gfp mask of the request.
But ALLOC_CMA is set in alloc_flags for all movable requests.
This can make zone_watermark_ok take incorrect decision.
Fix this by adding ALLOC_CMA only when request is of type
__GFP_CMA.

Change-Id: I15531968021c71cfbedd01d12307a203f0b408ca
Signed-off-by: Vinayak Menon <vinmenon@codeaurora.org>