sched: walt: Optimize task_util()
task_util() for WALT is currently defined as

p->ravg.demand / (sched_ravg_window >> SCHED_CAPACITY_SHIFT);

This math is required to scale the task demand to 1024 scale.
task_util() is used many times in task placement. So the calls
to this can be optimized by caching the scaled value when task
demand is calculated.

Change-Id: I0c170a10704ae3e8fe4e9f271e8e65c3923075e5
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: Yousef Algadri <yusufgadrie@gmail.com>
Signed-off-by: zeta96 [ Adapted ]